notebooks/pyproject.toml
[project]
name = "notebooks"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.39.0",
    "marimo>=0.9.16",
    "pillow>=11.0.0",
]
